requirements.txt
flask==3.1.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.10.0
opencv-python==4.13.0.92
opencv-contrib-python==4.13.0.92
pandas==3.0.1
openpyxl==3.1.5
reportlab==4.4.10
APScheduler==3.11.2
python-dotenv==1.2.2
requests==2.32.3
numpy==2.4.3
pillow==12.1.1
faiss-cpu==1.10.0
mediapipe==0.10.32